remove the post-install command and rely on detection at runtime to figure out which prebuilt binary to use
