--- npm/wizer/wizer
@@ -1,1 +1,0 @@
-placeholder file for the wizer binary to be copied to
